refactor(android): move the Twin algorithms out of this public repo
NameDetector, Sentiment and Prosody were written here. They should never have
been: on iOS the equivalents (HeuristicNER.swift, ProsodyFeatures.swift,
EntityGraph.swift) live in the private engine package and the public app carries
only UI, capture and persistence. The Android port did not follow that split.

They now live in github.com/intrepidkarthi/DailyVoxTwin under kotlin/engine, a
pure-JVM Gradle module the app includes by relative path — the Android analogue
of the SPM dependency iOS uses. The build FAILS with an explicit message when
the engine is absent rather than degrading, because an app with no name
detection and every mood at 0.00 looks like a working app until someone trusts
it.

What stayed here, because it plumbs rather than computes:
  AudioDecoder  — MediaCodec/MediaExtractor to PCM (iOS does this app-side too)
  Lexicon       — reads vader.txt out of assets and installs it into the engine

The rule is now written into ARCHITECTURE.md and android/README.md so it does
not depend on anyone remembering this conversation.

Tests moved with the code they cover: the engine's 26 tests run in its own repo
with no Android SDK. What stayed is LexiconAssetTest, which is the one thing
only this repo can check — that the real 7,517-entry table is actually present
and shippable.

app: 12 tests, 0 failures. Engine: 26 tests, 0 failures. Debug and release
builds verified against the private module.

NOTE: NameDetector.kt and Sentiment.kt were pushed to this public repo in
earlier commits and remain in its git HISTORY. Removing them from HEAD does not
remove them from history, and anyone who cloned already has them. Scrubbing
history means a force-push that breaks every existing clone — that is a call for
the repo owner, not something to do silently.
